requirements.txt
kserve == 0.7.0
ultralytics
tensorflow
tensorboard
cmake 
onnx
pyyaml>=5.3.1
Pillow
mss
timm
dataclasses
ftfy
regex
fasttext
scikit-learn
lvis
nltk
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/detectron2.git
fiftyone
filterpy
patchify
mpi4py
json_tricks
yacs
pandas
timm==0.4.12
numpy==1.23.5
einops
fvcore
transformers==4.19.2
sentencepiece
regex
nltk
vision-datasets==0.2.2
pycocotools==2.0.4
diffdist
pyarrow
cityscapesscripts
shapely
scikit-image
mup
gradio==3.13.0
scann
kornia==0.6.4
torchmetrics==0.6.0

#MQTT
paho-mqtt==1.6.1
piexif

